remove register page 4 (review info page)
